+++ etc/scripts/init.sh
@@ -12,9 +12,13 @@
 
 # Script to execute when starting
 SCRIPT="/usr/bin/chronograf"
+export HOST="0.0.0.0"
+export PORT="8888"
+export BOLT_PATH="/var/lib/chronograf/chronograf-v1.db"
+export CANNED_PATH="/usr/share/chronograf/canned"
 # Options to pass to the script on startup
 . /etc/default/chronograf
-SCRIPT_OPTS="--host 0.0.0.0 --port 8888 -b /var/lib/chronograf/chronograf-v1.db -c /usr/share/chronograf/canned ${CHRONOGRAF_OPTS}"
+SCRIPT_OPTS="${CHRONOGRAF_OPTS}"
 
 # User to run the process under
 RUNAS=chronograf