+++ ui/src/admin/components/chronograf/OrganizationsTable.js
@@ -2,12 +2,9 @@ import React, {Component, PropTypes} from 'react'
 
 import uuid from 'node-uuid'
 
-import Authorized, {SUPERADMIN_ROLE} from 'src/auth/Authorized'
-
 import OrganizationsTableRow from 'src/admin/components/chronograf/OrganizationsTableRow'
 import OrganizationsTableRowNew from 'src/admin/components/chronograf/OrganizationsTableRowNew'
 import QuestionMarkTooltip from 'shared/components/QuestionMarkTooltip'
-import SlideToggle from 'shared/components/SlideToggle'
 
 import {PUBLIC_TOOLTIP} from 'src/admin/constants/index'
 
@@ -42,8 +39,6 @@ class OrganizationsTable extends Component {
       onChooseDefaultRole,
       onTogglePublic,
       currentOrganization,
-      authConfig: {superAdminNewUsers},
-      onChangeAuthConfig,
     } = this.props
     const {isCreatingOrganization} = this.state
 
@@ -52,6 +47,15 @@ class OrganizationsTable extends Component {
       ? ''
       : 's'}`
 
+    if (!organizations.length) {
+      return (
+        <div className="panel panel-default">
+          <div className="panel-body">
+            <div className="page-spinner" />
+          </div>
+        </div>
+      )
+    }
     return (
       <div className="panel panel-default">
         <div className="panel-heading u-flex u-ai-center u-jc-space-between">
@@ -94,35 +98,13 @@ class OrganizationsTable extends Component {
               currentOrganization={currentOrganization}
             />
           )}
-          <Authorized requiredRole={SUPERADMIN_ROLE}>
-            <table className="table v-center superadmin-config">
-              <thead>
-                <tr>
-                  <th style={{width: 70}}>Config</th>
-                  <th />
-                </tr>
-              </thead>
-              <tbody>
-                <tr>
-                  <td style={{width: 70}}>
-                    <SlideToggle
-                      size="xs"
-                      active={superAdminNewUsers}
-                      onToggle={onChangeAuthConfig('superAdminNewUsers')}
-                    />
-                  </td>
-                  <td>All new users are SuperAdmins</td>
-                </tr>
-              </tbody>
-            </table>
-          </Authorized>
         </div>
       </div>
     )
   }
 }
 
-const {arrayOf, bool, func, shape, string} = PropTypes
+const {arrayOf, func, shape, string} = PropTypes
 
 OrganizationsTable.propTypes = {
   organizations: arrayOf(
@@ -140,9 +122,5 @@ OrganizationsTable.propTypes = {
   onRenameOrg: func.isRequired,
   onTogglePublic: func.isRequired,
   onChooseDefaultRole: func.isRequired,
-  onChangeAuthConfig: func.isRequired,
-  authConfig: shape({
-    superAdminNewUsers: bool,
-  }),
 }
 export default OrganizationsTable